Create dsse module and move predicate models to it
