+++ sigstore/dsse/_predicate.py
